rename scene
