+++ include/ramiel/ramiel.h
@@ -1,4 +1,4 @@
 #pragma once
 
-#include "../src/scene.h"
+#include "../src/ramiel.h"
 #include "../src/graphics/vertexshader.h"
